Move GeoColumn to its own file.
